@@ -26,7 +26,8 @@
         "docs": "https://github.com/getgrav/grav-plugin-login/blob/master/README.md"
     },
     "require": {
-        "php": ">=5.5.9",
+        "php": ">=7.1.3",
+        "ext-json": "*",
         "birke/rememberme": "1.*",
         "robthree/twofactorauth": "^1.6",
         "bacon/bacon-qr-code": "^1.0"
@@ -34,6 +35,11 @@
     "autoload": {
         "psr-4": {
             "Grav\\Plugin\\Login\\": "classes/"
+        }
+    },
+    "config": {
+        "platform": {
+            "php": "7.1.3"
         }
     }
 }
