@@ -1,16 +1,23 @@
 <?php
 namespace Grav\Plugin;
 
+use Composer\Autoload\ClassLoader;
+use Grav\Common\Cache;
+use Grav\Common\Debugger;
 use Grav\Common\File\CompiledYamlFile;
 use Grav\Common\Grav;
+use Grav\Common\Helpers\LogViewer;
 use Grav\Common\Inflector;
 use Grav\Common\Language\Language;
+use Grav\Common\Page\Interfaces\PageInterface;
 use Grav\Common\Page\Page;
 use Grav\Common\Page\Pages;
 use Grav\Common\Plugin;
+use Grav\Common\Session;
 use Grav\Common\Uri;
+use Grav\Common\User\Interfaces\UserCollectionInterface;
 use Grav\Common\Utils;
-use Grav\Common\User\User;
+use Grav\Framework\Session\Exceptions\SessionException;
 use Grav\Plugin\Admin\Admin;
 use Grav\Plugin\Admin\Popularity;
 use Grav\Plugin\Admin\Themes;
@@ -19,7 +26,6 @@ use Grav\Plugin\Admin\Twig\AdminTwigExtension;
 use Grav\Plugin\Form\Form;
 use Grav\Plugin\Login\Login;
 use RocketTheme\Toolbox\Event\Event;
-use RocketTheme\Toolbox\Session\Session;
 
 class AdminPlugin extends Plugin
 {
@@ -27,56 +33,37 @@ class AdminPlugin extends Plugin
         'blueprints' => 1000,
     ];
 
-    /**
-     * @var bool
-     */
+    /** @var bool */
     protected $active = false;
 
-    /**
-     * @var string
-     */
+    /** @var string */
     protected $template;
 
-    /**
-     * @var  string
-     */
+    /** @var  string */
     protected $theme;
 
-    /**
-     * @var string
-     */
+    /** @var string */
     protected $route;
 
-    /**
-     * @var string
-     */
+    /** @var string */
     protected $admin_route;
 
-    /**
-     * @var Uri
-     */
+    /** @var Uri */
     protected $uri;
 
-    /**
-     * @var Admin
-     */
+    /** @var Admin */
     protected $admin;
 
-    /**
-     * @var Session
-     */
+    /** @var Session */
     protected $session;
 
-    /**
-     * @var Popularity
-     */
+    /** @var Popularity */
     protected $popularity;
 
-    /**
-     * @var string
-     */
+    /** @var string */
     protected $base;
 
+    /** @var string */
     protected $version;
 
     /**
@@ -84,45 +71,67 @@ class AdminPlugin extends Plugin
      */
     public static function getSubscribedEvents()
     {
-
         return [
             'onPluginsInitialized' => [
-                                        ['setup', 100000],
-                                        ['onPluginsInitialized', 1001]
-                                      ],
+                ['autoload', 100001],
+                ['setup', 100000],
+                ['onPluginsInitialized', 1001]
+              ],
             'onPageInitialized'    => ['onPageInitialized', 0],
-            'onShutdown'           => ['onShutdown', 1000],
             'onFormProcessed'      => ['onFormProcessed', 0],
+            'onShutdown'           => ['onShutdown', 1000],
             'onAdminDashboard'     => ['onAdminDashboard', 0],
             'onAdminTools'         => ['onAdminTools', 0],
         ];
-
-    }
-
-    public function onPageInitialized()
-    {
-        $page = $this->grav['page'];
-
-        $template = $this->grav['uri']->param('tmpl');
-
-        if ($template) {
-            $page->template($template);
-        }
     }
 
     /**
+     * Get list of form field types specified in this plugin. Only special types needs to be listed.
+     *
+     * @return array
+     */
+    public function getFormFieldTypes()
+    {
+        return [
+            'column'   => [
+                'input@' => false
+            ],
+            'columns'  => [
+                'input@' => false
+            ],
+            'fieldset' => [
+                'input@' => false
+            ],
+            'section'  => [
+                'input@' => false
+            ],
+            'list'     => [
+                'array' => true
+            ],
+            'file'     => [
+                'array' => true
+            ]
+        ];
+    }
+
+    /**
+     * [onPluginsInitialized:100000] Composer autoload.
+     *
+     * @return ClassLoader
+     */
+    public function autoload()
+    {
+        return require __DIR__ . '/vendor/autoload.php';
+    }
+
+    /**
+     * [onPluginsInitialized:100000]
+     *
      * If the admin path matches, initialize the Login plugin configuration and set the admin
      * as active.
      */
     public function setup()
     {
-        // Autoloader
-        spl_autoload_register(function ($class) {
-            if (Utils::startsWith($class, 'Grav\Plugin\Admin')) {
-                require_once __DIR__ .'/classes/' . strtolower(basename(str_replace("\\", '/', $class))) . '.php';
-            }
-        });
-
         $route = $this->config->get('plugins.admin.route');
         if (!$route) {
             return;
@@ -144,38 +153,94 @@ class AdminPlugin extends Plugin
 
         // Only activate admin if we're inside the admin path.
         if ($this->isAdminPath()) {
-            $this->grav['session']->init();
+            try {
+                $this->grav['session']->init();
+            } catch (SessionException $e) {
+                $this->grav['session']->init();
+                $message = 'Session corruption detected, restarting session...';
+
+                /** @var Debugger $debugger */
+                $debugger = $this->grav['debugger'];
+                $debugger->addMessage($message);
+
+                $this->grav['messages']->add($message, 'error');
+            }
             $this->active = true;
 
             // Set cache based on admin_cache option
-            if (method_exists($this->grav['cache'], 'setEnabled')) {
-                $this->grav['cache']->setEnabled($this->config->get('plugins.admin.cache_enabled'));
+            $this->grav['cache']->setEnabled($this->config->get('plugins.admin.cache_enabled'));
+            $pages = $this->grav['pages'];
+            if (method_exists($pages, 'setCheckMethod')) {
+                // Force file hash checks to fix caching on moved and deleted pages.
+                $pages->setCheckMethod('hash');
             }
         }
     }
 
     /**
-     * Validate a value. Currently validates
+     * [onPluginsInitialized:1001]
      *
-     * - 'user' for username format and username availability.
-     * - 'password1' for password format
-     * - 'password2' for equality to password1
-     *
-     * @param string $type  The field type
-     * @param string $value The field value
-     * @param string $extra Any extra value required
-     *
-     * @return bool
+     * If the admin plugin is set as active, initialize the admin
      */
-    protected function validate($type, $value, $extra = '')
+    public function onPluginsInitialized()
     {
-        /** @var Login $login */
-        $login = $this->grav['login'];
+        // Only activate admin if we're inside the admin path.
+        if ($this->active) {
+            // Store this version.
+            $this->version = $this->getBlueprint()->get('version');
 
-        return $login->validateField($type, $value, $extra);
+            // Have a unique Admin-only Cache key
+            if (method_exists($this->grav['cache'], 'setKey')) {
+                /** @var Cache $cache */
+                $cache = $this->grav['cache'];
+                $cache_key = $cache->getKey();
+                $cache->setKey($cache_key . '$');
+            }
+
+            // Turn on Twig autoescaping
+            if (method_exists($this->grav['twig'], 'setAutoescape') && $this->grav['uri']->param('task') !== 'processmarkdown') {
+                $this->grav['twig']->setAutoescape(true);
+            }
+
+            $this->grav['debugger']->addMessage('Admin v' . $this->version);
+            $this->initializeAdmin();
+
+            // Disable Asset pipelining (old method - remove this after Grav is updated)
+            if (!method_exists($this->grav['assets'], 'setJsPipeline')) {
+                $this->config->set('system.assets.css_pipeline', false);
+                $this->config->set('system.assets.js_pipeline', false);
+            }
+
+            // Replace themes service with admin.
+            $this->grav['themes'] = function () {
+                return new Themes($this->grav);
+            };
+        }
+
+        // We need popularity no matter what
+        $this->popularity = new Popularity();
+
+        // Fire even to register permissions from other plugins
+        $this->grav->fireEvent('onAdminRegisterPermissions', new Event(['admin' => $this->admin]));
     }
 
     /**
+     * [onPageInitialized:0]
+     */
+    public function onPageInitialized()
+    {
+        $page = $this->grav['page'];
+
+        $template = $this->grav['uri']->param('tmpl');
+
+        if ($template) {
+            $page->template($template);
+        }
+    }
+
+    /**
+     * [onFormProcessed:0]
+     *
      * Process the admin registration form.
      *
      * @param Event $event
@@ -232,18 +297,18 @@ class AdminPlugin extends Plugin
 
                 $inflector = new Inflector();
 
-                $data['fullname'] = isset($data['fullname']) ? $data['fullname'] : $inflector->titleize($username);
-                $data['title'] = isset($data['title']) ? $data['title'] : 'Administrator';
+                $data['fullname'] = $data['fullname'] ?? $inflector->titleize($username);
+                $data['title'] = $data['title'] ?? 'Administrator';
                 $data['state'] = 'enabled';
                 $data['access'] = ['admin' => ['login' => true, 'super' => true], 'site' => ['login' => true]];
 
+                /** @var UserCollectionInterface $users */
+                $users = $this->grav['accounts'];
+
                 // Create user object and save it
-                $user = new User($data);
-                $file = CompiledYamlFile::instance($this->grav['locator']->findResource('user://accounts/' . $username . YAML_EXT,
-                    true, true));
-                $user->file($file);
+                $user = $users->load($username);
+                $user->update($data);
                 $user->save();
-                $user = User::load($username);
 
                 //Login user
                 $this->grav['session']->user = $user;
@@ -261,55 +326,56 @@ class AdminPlugin extends Plugin
     }
 
     /**
-     * If the admin plugin is set as active, initialize the admin
+     * [onShutdown:1000]
+     *
+     * Handles the shutdown
      */
-    public function onPluginsInitialized()
+    public function onShutdown()
     {
-        // Only activate admin if we're inside the admin path.
         if ($this->active) {
-            // Store this version.
-            $this->version = $this->getBlueprint()->version;
-
-            // Have a unique Admin-only Cache key
-            if (method_exists($this->grav['cache'], 'setKey')) {
-                $cache = $this->grav['cache'];
-                $cache_key = $cache->getKey();
-                $cache->setKey($cache_key . '$');
+            //only activate when Admin is active
+            if ($this->admin->shouldLoadAdditionalFilesInBackground()) {
+                $this->admin->loadAdditionalFilesInBackground();
             }
-
-            // Turn on Twig autoescaping
-            if (method_exists($this->grav['twig'], 'setAutoescape') && $this->grav['uri']->param('task') !== 'processmarkdown') {
-                $this->grav['twig']->setAutoescape(true);
+        } else {
+            //if popularity is enabled, track non-admin hits
+            if ($this->config->get('plugins.admin.popularity.enabled')) {
+                $this->popularity->trackHit();
             }
-
-            $this->grav['debugger']->addMessage("Admin Basic");
-            $this->initializeAdmin();
-
-            // Disable Asset pipelining (old method - remove this after Grav is updated)
-            if (!method_exists($this->grav['assets'], 'setJsPipeline')) {
-                $this->config->set('system.assets.css_pipeline', false);
-                $this->config->set('system.assets.js_pipeline', false);
-            }
-
-            // Replace themes service with admin.
-            $this->grav['themes'] = function () {
-                return new Themes($this->grav);
-            };
         }
-
-        // We need popularity no matter what
-        $this->popularity = new Popularity();
-
-        // Fire even to register permissions from other plugins
-        $this->grav->fireEvent('onAdminRegisterPermissions', new Event(['admin' => $this->admin]));
     }
 
-    protected function initializeController($task, $post)
+    /**
+     * [onAdminDashboard:0]
+     */
+    public function onAdminDashboard()
     {
-        $controller = new AdminController();
-        $controller->initialize($this->grav, $this->template, $task, $this->route, $post);
-        $controller->execute();
-        $controller->redirect();
+        $this->grav['twig']->plugins_hooked_dashboard_widgets_top[] = ['template' => 'dashboard-maintenance'];
+        $this->grav['twig']->plugins_hooked_dashboard_widgets_top[] = ['template' => 'dashboard-statistics'];
+        $this->grav['twig']->plugins_hooked_dashboard_widgets_top[] = ['template' => 'dashboard-notifications'];
+        $this->grav['twig']->plugins_hooked_dashboard_widgets_top[] = ['template' => 'dashboard-feed'];
+        $this->grav['twig']->plugins_hooked_dashboard_widgets_main[] = ['template' => 'dashboard-pages'];
+    }
+
+
+    /**
+     * [onAdminTools:0]
+     *
+     * Provide the tools for the Tools page, currently only direct install
+     *
+     * @return Event
+     */
+    public function onAdminTools(Event $event)
+    {
+        $event['tools'] = array_merge($event['tools'], [
+            'backups'        => [['admin.maintenance', 'admin.super'], $this->grav['language']->translate('PLUGIN_ADMIN.BACKUPS')],
+            'scheduler'      => [['admin.super'], $this->grav['language']->translate('PLUGIN_ADMIN.SCHEDULER')],
+            'logs'           => [['admin.super'], $this->grav['language']->translate('PLUGIN_ADMIN.LOGS')],
+            'reports'        => [['admin.super'], $this->grav['language']->translate('PLUGIN_ADMIN.REPORTS')],
+            'direct-install' => [['admin.super'], $this->grav['language']->translate('PLUGIN_ADMIN.DIRECT_INSTALL')],
+        ]);
+
+        return $event;
     }
 
     /**
@@ -332,14 +398,14 @@ class AdminPlugin extends Plugin
         // Set original route for the home page.
         $home = '/' . trim($this->config->get('system.home.alias'), '/');
 
-        // set the default if not set before
-        $this->session->expert = $this->session->expert ?: false;
-
         // set session variable if it's passed via the url
         if ($this->uri->param('mode') === 'expert') {
             $this->session->expert = true;
         } elseif ($this->uri->param('mode') === 'normal') {
             $this->session->expert = false;
+        } else {
+            // set the default if not set before
+            $this->session->expert = $this->session->expert ?? false;
         }
 
         /** @var Pages $pages */
@@ -366,7 +432,7 @@ class AdminPlugin extends Plugin
         Pages::types();
 
         // Handle tasks.
-        $this->admin->task = $task = !empty($post['task']) ? $post['task'] : $this->uri->param('task');
+        $this->admin->task = $task = $this->grav['task'];
         if ($task) {
             $this->initializeController($task, $post);
         } elseif ($this->template === 'logs' && $this->route) {
@@ -384,7 +450,7 @@ class AdminPlugin extends Plugin
 
         // Replace page service with admin.
         $this->grav['page'] = function () use ($self) {
-            $page = new Page;
+            $page = new Page();
             $page->expires(0);
 
             if ($this->grav['user']->authorize('admin.login')) {
@@ -413,7 +479,7 @@ class AdminPlugin extends Plugin
                     continue;
                 }
 
-                $path = $locator->findResource("user://plugins/{$plugin->name}/admin/pages/{$self->template}.md");
+                $path = $locator->findResource("plugins://{$plugin->name}/admin/pages/{$self->template}.md");
 
                 if ($path) {
                     $page->init(new \SplFileInfo($path));
@@ -428,16 +494,30 @@ class AdminPlugin extends Plugin
 
         if (empty($this->grav['page'])) {
             if ($this->grav['user']->authenticated) {
-                $event = $this->grav->fireEvent('onPageNotFound');
+                $event = new Event(['page' => null]);
+                $event->page = null;
+                $event = $this->grav->fireEvent('onPageNotFound', $event);
+                /** @var PageInterface $page */
+                $page = $event->page;
 
-                if (isset($event->page)) {
-                    unset($this->grav['page']);
-                    $this->grav['page'] = $event->page;
-                } else {
-                    throw new \RuntimeException('Page Not Found', 404);
+                if (!$page || !$page->routable()) {
+                    $error_file = $this->grav['locator']->findResource('plugins://admin/pages/admin/error.md');
+                    $page = new Page();
+                    $page->init(new \SplFileInfo($error_file));
+                    $page->slug(basename($this->route));
+                    $page->routable(true);
                 }
+
+                unset($this->grav['page']);
+                $this->grav['page'] = $page;
             } else {
-                $this->grav->redirect($this->admin_route);
+                // Not Found and not logged in: Display login page.
+                $login_file = $this->grav['locator']->findResource('plugins://admin/pages/admin/login.md');
+                $page = new Page();
+                $page->init(new \SplFileInfo($login_file));
+                $page->slug(basename($this->route));
+                unset($this->grav['page']);
+                $this->grav['page'] = $page;
             }
         }
 
@@ -480,16 +560,20 @@ class AdminPlugin extends Plugin
         $twig->twig_vars['location'] = $this->template;
         $twig->twig_vars['base_url_relative_frontend'] = $twig->twig_vars['base_url_relative'] ?: '/';
         $twig->twig_vars['admin_route'] = trim($this->admin_route, '/');
+        $twig->twig_vars['template_route'] = $this->template;
         $twig->twig_vars['current_route'] = '/' . $twig->twig_vars['admin_route'] . '/' . $this->template . '/' . $this->route;
         $twig->twig_vars['base_url_relative'] = $twig->twig_vars['base_url_simple'] . '/' . $twig->twig_vars['admin_route'];
         $twig->twig_vars['current_url'] = rtrim($twig->twig_vars['base_url_relative'] . '/' . $this->template . '/' . $this->route, '/');
         $theme_url = '/' . ltrim($this->grav['locator']->findResource('plugin://admin/themes/' . $this->theme,
             false), '/');
         $twig->twig_vars['theme_url'] = $theme_url;
+        $twig->twig_vars['preset_url'] = $twig->twig_vars['preset_url'] ?? $theme_url;
         $twig->twig_vars['base_url'] = $twig->twig_vars['base_url_relative'];
         $twig->twig_vars['base_path'] = GRAV_ROOT;
         $twig->twig_vars['admin'] = $this->admin;
         $twig->twig_vars['admin_version'] = $this->version;
+        $twig->twig_vars['logviewer'] = new LogViewer();
+        $twig->twig_vars['form_max_filesize'] = Utils::getUploadLimit() / 1024 / 1024;
 
         $fa_icons_file = CompiledYamlFile::instance($this->grav['locator']->findResource('plugin://admin/themes/grav/templates/forms/fields/iconpicker/icons' . YAML_EXT));
         $fa_icons = $fa_icons_file->content();
@@ -538,60 +622,195 @@ class AdminPlugin extends Plugin
         }
     }
 
-    /**
-     * Handles the shutdown
-     */
-    public function onShutdown()
+    // Add images to twig template paths to allow inclusion of SVG files
+    public function onTwigLoader()
     {
-        if ($this->active) {
-            //only activate when Admin is active
-            if ($this->admin->shouldLoadAdditionalFilesInBackground()) {
-                $this->admin->loadAdditionalFilesInBackground();
-            }
-        } else {
-            //if popularity is enabled, track non-admin hits
-            if ($this->config->get('plugins.admin.popularity.enabled')) {
-                $this->popularity->trackHit();
+        $theme_paths = Grav::instance()['locator']->findResources('plugins://admin/themes/' . $this->theme . '/images');
+        foreach($theme_paths as $images_path) {
+            $this->grav['twig']->addPath($images_path, 'admin-images');
+        }
+    }
+
+    /**
+     * Add the Admin Twig Extensions
+     */
+    public function onTwigExtensions()
+    {
+        require_once __DIR__ . '/classes/Twig/AdminTwigExtension.php';
+
+        $this->grav['twig']->twig->addExtension(new AdminTwigExtension);
+    }
+
+    public function onAdminAfterSave(Event $event)
+    {
+        // Special case to redirect after changing the admin route to avoid 'breaking'
+        $obj = $event['object'];
+
+        if (null !== $obj && method_exists($obj, 'blueprints')) {
+            $blueprint = $obj->blueprints()->getFilename();
+
+            if ($blueprint === 'admin/blueprints' && isset($obj->route) && $this->admin_route !== $obj->route) {
+                $redirect = preg_replace('/^' . str_replace('/','\/',$this->admin_route) . '/',$obj->route,$this->uri->path());
+                $this->grav->redirect($redirect);
             }
         }
     }
 
     /**
-     * Get list of form field types specified in this plugin. Only special types needs to be listed.
+     * Convert some types where we want to process out of the standard config path
      *
-     * @return array
+     * @param Event $e
      */
-    public function getFormFieldTypes()
+    public function onAdminData(Event $e)
     {
-        return [
-            'column'   => [
-                'input@' => false
-            ],
-            'columns'  => [
-                'input@' => false
-            ],
-            'fieldset' => [
-                'input@' => false
-            ],
-            'section'  => [
-                'input@' => false
-            ],
-            'tab'      => [
-                'input@' => false
-            ],
-            'tabs'     => [
-                'input@' => false
-            ],
-            'key'      => [
-                'input@' => false
-            ],
-            'list'     => [
-                'array' => true
-            ],
-            'file'     => [
-                'array' => true
-            ]
+        $type = $e['type'] ?? null;
+        switch ($type) {
+            case 'tools/scheduler':
+                $e['type'] = 'config/scheduler';
+                break;
+            case  'tools':
+            case 'tools/backups':
+                $e['type'] = 'config/backups';
+                break;
+        }
+    }
+
+    public function onOutputGenerated()
+    {
+        // Clear flash objects for previously uploaded files whenever the user switches page or reloads
+        // ignoring any JSON / extension call
+        if ($this->admin->task !== 'save' && empty($this->uri->extension())) {
+            // Discard any previously uploaded files session and remove all uploaded files.
+            if ($flash = $this->session->getFlashObject('files-upload')) {
+                $flash = new \RecursiveIteratorIterator(new \RecursiveArrayIterator($flash));
+                foreach ($flash as $key => $value) {
+                    if ($key !== 'tmp_name') {
+                        continue;
+                    }
+                    @unlink($value);
+                }
+            }
+        }
+    }
+
+    /**
+     * Initial stab at registering permissions (WIP)
+     *
+     * @param Event $e
+     */
+    public function onAdminRegisterPermissions(Event $e)
+    {
+        $admin = $e['admin'];
+        $permissions = [
+            'admin.super'         => 'boolean',
+            'admin.login'         => 'boolean',
+            'admin.cache'         => 'boolean',
+            'admin.configuration' => 'boolean',
+            'admin.configuration_system' => 'boolean',
+            'admin.configuration_site' => 'boolean',
+            'admin.configuration_media' => 'boolean',
+            'admin.configuration_info' => 'boolean',
+            'admin.settings'      => 'boolean',
+            'admin.pages'         => 'boolean',
+            'admin.maintenance'   => 'boolean',
+            'admin.statistics'    => 'boolean',
+            'admin.plugins'       => 'boolean',
+            'admin.themes'        => 'boolean',
+            'admin.tools'         => 'boolean',
+            'admin.users'         => 'boolean',
         ];
+        $admin->addPermissions($permissions);
+    }
+
+    /**
+     * Check if the current route is under the admin path
+     *
+     * @return bool
+     */
+    public function isAdminPath()
+    {
+        $route = $this->uri->route();
+
+        return $route === $this->base || 0 === strpos($route, $this->base . '/');
+    }
+
+    /**
+     * Helper function to replace Pages::Types()
+     * and to provide an event to manipulate the data
+     *
+     * Dispatches 'onAdminPageTypes' event
+     * with 'types' data member which is a
+     * reference to the data
+     */
+    public static function pagesTypes()
+    {
+        $types = Pages::types();
+
+        // First filter by configuration
+        $hideTypes = Grav::instance()['config']->get('plugins.admin.hide_page_types', []);
+        foreach ((array) $hideTypes as $type) {
+            unset($types[$type]);
+        }
+
+        // Allow manipulating of the data by event
+        $e = new Event(['types' => &$types]);
+        Grav::instance()->fireEvent('onAdminPageTypes', $e);
+
+        return $types;
+    }
+
+    /**
+     * Helper function to replace Pages::modularTypes()
+     * and to provide an event to manipulate the data
+     *
+     * Dispatches 'onAdminModularPageTypes' event
+     * with 'types' data member which is a
+     * reference to the data
+     */
+    public static function pagesModularTypes()
+    {
+        $types = Pages::modularTypes();
+
+        // First filter by configuration
+        $hideTypes = (array) Grav::instance()['config']->get('plugins.admin.hide_modular_page_types', []);
+        foreach ($hideTypes as $type) {
+            unset($types[$type]);
+        }
+
+        // Allow manipulating of the data by event
+        $e = new Event(['types' => &$types]);
+        Grav::instance()->fireEvent('onAdminModularPageTypes', $e);
+
+        return $types;
+    }
+
+    /**
+     * Validate a value. Currently validates
+     *
+     * - 'user' for username format and username availability.
+     * - 'password1' for password format
+     * - 'password2' for equality to password1
+     *
+     * @param string $type  The field type
+     * @param string $value The field value
+     * @param string $extra Any extra value required
+     *
+     * @return bool
+     */
+    protected function validate($type, $value, $extra = '')
+    {
+        /** @var Login $login */
+        $login = $this->grav['login'];
+
+        return $login->validateField($type, $value, $extra);
+    }
+
+    protected function initializeController($task, $post)
+    {
+        $controller = new AdminController();
+        $controller->initialize($this->grav, $this->template, $task, $this->route, $post);
+        $controller->execute();
+        $controller->redirect();
     }
 
     /**
@@ -604,12 +823,14 @@ class AdminPlugin extends Plugin
         $this->enable([
             'onTwigExtensions'           => ['onTwigExtensions', 1000],
             'onPagesInitialized'         => ['onPagesInitialized', 1000],
+            'onTwigLoader'               => ['onTwigLoader', 1000],
             'onTwigTemplatePaths'        => ['onTwigTemplatePaths', 1000],
             'onTwigSiteVariables'        => ['onTwigSiteVariables', 1000],
             'onAssetsInitialized'        => ['onAssetsInitialized', 1000],
             'onAdminRegisterPermissions' => ['onAdminRegisterPermissions', 0],
             'onOutputGenerated'          => ['onOutputGenerated', 0],
             'onAdminAfterSave'           => ['onAdminAfterSave', 0],
+            'onAdminData'                => ['onAdminData', 0],
         ]);
 
         // Autoload classes
@@ -642,16 +863,12 @@ class AdminPlugin extends Plugin
             $this->route = array_shift($array);
         }
 
-        // Initialize admin class.
+        // Initialize admin class (also registers it to Grav services).
         $this->admin = new Admin($this->grav, $this->admin_route, $this->template, $this->route);
-
-
-        // And store the class into DI container.
-        $this->grav['admin'] = $this->admin;
 
         // Double check we have system.yaml, site.yaml etc
         $config_path = $this->grav['locator']->findResource('user://config');
-        foreach ($this->admin->configurations() as $config_file) {
+        foreach ($this->admin::configurations() as $config_file) {
             $config_file = "{$config_path}/{$config_file}.yaml";
             if (!file_exists($config_file)) {
                 touch($config_file);
@@ -742,7 +959,7 @@ class AdminPlugin extends Plugin
 
         foreach ($strings as $string) {
             $separator = (end($strings) === $string) ? '' : ',';
-            $translations .= '"' . $string . '": "' . htmlspecialchars($this->admin->translate('PLUGIN_ADMIN.' . $string)) . '"' . $separator;
+            $translations .= '"' . $string . '": "' . htmlspecialchars($this->admin::translate('PLUGIN_ADMIN.' . $string)) . '"' . $separator;
         }
 
         $translations .= '};';
@@ -751,170 +968,47 @@ class AdminPlugin extends Plugin
         $strings = ['RESOLUTION_MIN', 'RESOLUTION_MAX'];
         foreach ($strings as $string) {
             $separator = (end($strings) === $string) ? '' : ',';
-            $translations .= '"' . $string . '": "' . $this->admin->translate('PLUGIN_FORM.' . $string) . '"' . $separator;
+            $translations .= '"' . $string . '": "' . $this->admin::translate('PLUGIN_FORM.' . $string) . '"' . $separator;
         }
+        $translations .= '};';
 
+        $translations .= 'this.GravAdmin.translations.GRAV_CORE = {';
+        $strings = [
+            'NICETIME.SECOND',
+            'NICETIME.MINUTE',
+            'NICETIME.HOUR',
+            'NICETIME.DAY',
+            'NICETIME.WEEK',
+            'NICETIME.MONTH',
+            'NICETIME.YEAR',
+            'CRON.EVERY',
+            'CRON.EVERY_HOUR',
+            'CRON.EVERY_MINUTE',
+            'CRON.EVERY_DAY_OF_WEEK',
+            'CRON.EVERY_DAY_OF_MONTH',
+            'CRON.EVERY_MONTH',
+            'CRON.TEXT_PERIOD',
+            'CRON.TEXT_MINS',
+            'CRON.TEXT_TIME',
+            'CRON.TEXT_DOW',
+            'CRON.TEXT_MONTH',
+            'CRON.TEXT_DOM',
+            'CRON.ERROR1',
+            'CRON.ERROR2',
+            'CRON.ERROR3',
+            'CRON.ERROR4',
+            'MONTHS_OF_THE_YEAR',
+            'DAYS_OF_THE_WEEK'
+        ];
+        foreach ($strings as $string) {
+            $separator = (end($strings) === $string) ? '' : ',';
+            $translations .= '"' . $string . '": ' . json_encode($this->admin::translate('GRAV.'.$string)) . $separator;
+        }
         $translations .= '};';
 
         // set the actual translations state back
         $this->config->set('system.languages.translations', $translations_actual_state);
 
         $assets->addInlineJs($translations);
-    }
-
-    /**
-     * Add the Admin Twig Extensions
-     */
-    public function onTwigExtensions()
-    {
-        require_once __DIR__ . '/classes/Twig/AdminTwigExtension.php';
-
-        $this->grav['twig']->twig->addExtension(new AdminTwigExtension);
-    }
-
-    /**
-     * Check if the current route is under the admin path
-     *
-     * @return bool
-     */
-    public function isAdminPath()
-    {
-        $route = $this->uri->route();
-
-        return $route === $this->base || 0 === strpos($route, $this->base . '/');
-    }
-
-    public function onAdminAfterSave(Event $event)
-    {
-        // Special case to redirect after changing the admin route to avoid 'breaking'
-        $obj = $event['object'];
-
-        if (null !== $obj && method_exists($obj, 'blueprints')) {
-            $blueprint = $obj->blueprints()->getFilename();
-
-            if ($blueprint === 'admin/blueprints' && isset($obj->route) && $this->admin_route !== $obj->route) {
-                $redirect = preg_replace('/^' . str_replace('/','\/',$this->admin_route) . '/',$obj->route,$this->uri->path());
-                $this->grav->redirect($redirect);
-            }
-        }
-    }
-
-    /**
-     * Provide the tools for the Tools page, currently only direct install
-     *
-     * @return Event
-     */
-    public function onAdminTools(Event $event)
-    {
-        $event['tools'] = array_merge($event['tools'], [$this->grav['language']->translate('PLUGIN_ADMIN.DIRECT_INSTALL')]);
-        return $event;
-    }
-
-    public function onAdminDashboard()
-    {
-        $this->grav['twig']->plugins_hooked_dashboard_widgets_top[] = ['template' => 'dashboard-maintenance'];
-        $this->grav['twig']->plugins_hooked_dashboard_widgets_top[] = ['template' => 'dashboard-statistics'];
-        $this->grav['twig']->plugins_hooked_dashboard_widgets_top[] = ['template' => 'dashboard-notifications'];
-        $this->grav['twig']->plugins_hooked_dashboard_widgets_top[] = ['template' => 'dashboard-feed'];
-        $this->grav['twig']->plugins_hooked_dashboard_widgets_main[] = ['template' => 'dashboard-pages'];
-    }
-
-    public function onOutputGenerated()
-    {
-        // Clear flash objects for previously uploaded files
-        // whenever the user switches page / reloads
-        // ignoring any JSON / extension call
-        if ($this->admin->task !== 'save' && empty($this->uri->extension())) {
-            // Discard any previously uploaded files session.
-            // and if there were any uploaded file, remove them from the filesystem
-            if ($flash = $this->session->getFlashObject('files-upload')) {
-                $flash = new \RecursiveIteratorIterator(new \RecursiveArrayIterator($flash));
-                foreach ($flash as $key => $value) {
-                    if ($key !== 'tmp_name') {
-                        continue;
-                    }
-                    @unlink($value);
-                }
-            }
-        }
-    }
-
-    /**
-     * Initial stab at registering permissions (WIP)
-     *
-     * @param Event $e
-     */
-    public function onAdminRegisterPermissions(Event $e)
-    {
-        $admin = $e['admin'];
-        $permissions = [
-            'admin.super'         => 'boolean',
-            'admin.login'         => 'boolean',
-            'admin.cache'         => 'boolean',
-            'admin.configuration' => 'boolean',
-            'admin.configuration_system' => 'boolean',
-            'admin.configuration_site' => 'boolean',
-            'admin.configuration_media' => 'boolean',
-            'admin.configuration_info' => 'boolean',
-            'admin.settings'      => 'boolean',
-            'admin.pages'         => 'boolean',
-            'admin.maintenance'   => 'boolean',
-            'admin.statistics'    => 'boolean',
-            'admin.plugins'       => 'boolean',
-            'admin.themes'        => 'boolean',
-            'admin.tools'         => 'boolean',
-            'admin.users'         => 'boolean',
-        ];
-        $admin->addPermissions($permissions);
-    }
-
-    /**
-     * Helper function to replace Pages::Types()
-     * and to provide an event to manipulate the data
-     *
-     * Dispatches 'onAdminPageTypes' event
-     * with 'types' data member which is a
-     * reference to the data
-     */
-    public static function pagesTypes()
-    {
-        $types = Pages::types();
-
-        // First filter by configuration
-        $hideTypes = Grav::instance()['config']->get('plugins.admin.hide_page_types', []);
-        foreach ((array) $hideTypes as $type) {
-            unset($types[$type]);
-        }
-
-        // Allow manipulating of the data by event
-        $e = new Event(['types' => &$types]);
-        Grav::instance()->fireEvent('onAdminPageTypes', $e);
-
-        return $types;
-    }
-
-    /**
-     * Helper function to replace Pages::modularTypes()
-     * and to provide an event to manipulate the data
-     *
-     * Dispatches 'onAdminModularPageTypes' event
-     * with 'types' data member which is a
-     * reference to the data
-     */
-    public static function pagesModularTypes()
-    {
-        $types = Pages::modularTypes();
-
-        // First filter by configuration
-        $hideTypes = (array) Grav::instance()['config']->get('plugins.admin.hide_modular_page_types', []);
-        foreach ($hideTypes as $type) {
-            unset($types[$type]);
-        }
-
-        // Allow manipulating of the data by event
-        $e = new Event(['types' => &$types]);
-        Grav::instance()->fireEvent('onAdminModularPageTypes', $e);
-
-        return $types;
     }
 }
