@@ -1,22 +1,25 @@
 <?php
+
 /**
  * @package    Grav\Plugin\Login
  *
  * @copyright  Copyright (C) 2014 - 2017 RocketTheme, LLC. All rights reserved.
  * @license    MIT License; see LICENSE file for details.
  */
+
 namespace Grav\Plugin\Login;
 
 use Birke\Rememberme\Cookie;
 use Grav\Common\Config\Config;
 use Grav\Common\Data\Data;
 use Grav\Common\Grav;
-use Grav\Common\File\CompiledYamlFile;
 use Grav\Common\Language\Language;
-use Grav\Common\Page\Page;
+use Grav\Common\Page\Interfaces\PageInterface;
 use Grav\Common\Session;
-use Grav\Common\User\User;
+use Grav\Common\User\Interfaces\UserCollectionInterface;
+use Grav\Common\User\Interfaces\UserInterface;
 use Grav\Common\Uri;
+use Grav\Common\Utils;
 use Grav\Plugin\Email\Utils as EmailUtils;
 use Grav\Plugin\Login\Events\UserLoginEvent;
 use Grav\Plugin\Login\RememberMe\RememberMe;
@@ -76,7 +79,7 @@ class Login
      * @param array $credentials    Login credentials, eg: ['username' => '', 'password' => '']
      * @param array $options        Login options, eg: ['remember_me' => true]
      * @param array $extra          Example: ['authorize' => 'site.login', 'user' => null], undefined variables get set.
-     * @return User|UserLoginEvent  Returns event if $extra['return_event'] is true.
+     * @return UserInterface|UserLoginEvent  Returns event if $extra['return_event'] is true.
      */
     public function login(array $credentials, array $options = [], array $extra = [])
     {
@@ -133,15 +136,15 @@ class Login
     /**
      * Logout user.
      *
-     * @param array                 $options
-     * @param array|User            $extra      Array of: ['user' => $user, ...] or User object (deprecated).
-     * @return User|UserLoginEvent  Returns event if $extra['return_event'] is true.
+     * @param array                         $options
+     * @param array|UserInterface           $extra      Array of: ['user' => $user, ...] or UserInterface object (deprecated).
+     * @return UserInterface|UserLoginEvent Returns event if $extra['return_event'] is true.
      */
     public function logout(array $options = [], $extra = [])
     {
         $grav = Grav::instance();
 
-        if ($extra instanceof User) {
+        if ($extra instanceof UserInterface) {
             $extra = ['user' => $extra];
         } elseif (isset($extra['user'])) {
             $extra['user'] = $grav['user'];
@@ -170,6 +173,7 @@ class Login
      * @param array $options
      *
      * @return bool
+     * @deprecated Uses the Controller::taskLogin() event
      */
     public function authenticate($credentials, $options = ['remember_me' => true])
     {
@@ -206,16 +210,17 @@ class Login
      * Create a new user file
      *
      * @param array $data
+     * @param array $files
      *
-     * @return User
+     * @return UserInterface
      */
-    public function register($data)
+    public function register(array $data, array $files = [])
     {
         if (!isset($data['groups'])) {
             //Add new user ACL settings
             $groups = (array) $this->config->get('plugins.login.user_registration.groups', []);
 
-            if (count($groups) > 0) {
+            if (\count($groups) > 0) {
                 $data['groups'] = $groups;
             }
         }
@@ -223,19 +228,29 @@ class Login
         if (!isset($data['access'])) {
             $access = (array) $this->config->get('plugins.login.user_registration.access.site', []);
 
-            if (count($access) > 0) {
+            if (\count($access) > 0) {
                 $data['access']['site'] = $access;
             }
         }
 
         $username = $this->validateField('username', $data['username']);
 
-        $file = CompiledYamlFile::instance($this->grav['locator']->findResource('account://' . mb_strtolower($username) . YAML_EXT,
-            true, true));
+        /** @var UserCollectionInterface $users */
+        $users = $this->grav['accounts'];
 
         // Create user object and save it
-        $user = new User($data);
-        $user->file($file);
+        $user = $users->load($username);
+        if ($user->exists()) {
+            throw new \RuntimeException('User ' . $username . ' cannot be registered: user already exists!');
+        }
+
+        $user->update($data, $files);
+        if (isset($data['groups'])) {
+            $user->groups = $data['groups'];
+        }
+        if (isset($data['access'])) {
+            $user->access = $data['access'];
+        }
         $user->save();
 
         return $user;
@@ -275,28 +290,28 @@ class Login
                 break;
 
             case 'password2':
-                if (!is_string($value) || strcmp($value, $extra)) {
+                if (!\is_string($value) || strcmp($value, $extra)) {
                     throw new \RuntimeException('Passwords did not match.');
                 }
 
                 break;
 
             case 'email':
-                if (!is_string($value) || !filter_var($value, FILTER_VALIDATE_EMAIL)) {
+                if (!\is_string($value) || !filter_var($value, FILTER_VALIDATE_EMAIL)) {
                     throw new \RuntimeException('Not a valid email address');
                 }
 
                 break;
 
             case 'permissions':
-                if (!is_string($value) || !in_array($value, ['a', 's', 'b'])) {
+                if (!\is_string($value) || !\in_array($value, ['a', 's', 'b'], true)) {
                     throw new \RuntimeException('Permissions ' . $value . ' are invalid.');
                 }
 
                 break;
 
             case 'fullname':
-                if (!is_string($value) || trim($value) === '') {
+                if (!\is_string($value) || trim($value) === '') {
                     throw new \RuntimeException('Fullname cannot be empty');
                 }
 
@@ -317,12 +332,12 @@ class Login
     /**
      * Handle the email to notify the user account creation to the site admin.
      *
-     * @param User $user
+     * @param UserInterface $user
      *
      * @return bool True if the action was performed.
      * @throws \RuntimeException
      */
-    public function sendNotificationEmail(User $user)
+    public function sendNotificationEmail(UserInterface $user)
     {
         if (empty($user->email)) {
             throw new \RuntimeException($this->language->translate('PLUGIN_LOGIN.USER_NEEDS_EMAIL_FIELD'));
@@ -356,12 +371,12 @@ class Login
     /**
      * Handle the email to welcome the new user
      *
-     * @param User $user
+     * @param UserInterface $user
      *
      * @return bool True if the action was performed.
      * @throws \RuntimeException
      */
-    public function sendWelcomeEmail(User $user)
+    public function sendWelcomeEmail(UserInterface $user)
     {
         if (empty($user->email)) {
             throw new \RuntimeException($this->language->translate('PLUGIN_LOGIN.USER_NEEDS_EMAIL_FIELD'));
@@ -392,12 +407,12 @@ class Login
     /**
      * Handle the email to activate the user account.
      *
-     * @param User $user
+     * @param UserInterface $user
      *
      * @return bool True if the action was performed.
      * @throws \RuntimeException
      */
-    public function sendActivationEmail(User $user)
+    public function sendActivationEmail(UserInterface $user)
     {
         if (empty($user->email)) {
             throw new \RuntimeException($this->language->translate('PLUGIN_LOGIN.USER_NEEDS_EMAIL_FIELD'));
@@ -453,14 +468,14 @@ class Login
             $timeout = $config->get('plugins.login.rememberme.timeout');
 
             // Setup storage for RememberMe cookies
-            $storage = new TokenStorage('user://data/rememberme', $timeout);
+            $storage = new TokenStorage('user-data://rememberme', $timeout);
             $this->rememberMe = new RememberMe($storage);
             $this->rememberMe->setCookieName($cookieName);
             $this->rememberMe->setExpireTime($timeout);
 
             // Hardening cookies with user-agent and random salt or
             // fallback to use system based cache key
-            $server_agent = isset($_SERVER['HTTP_USER_AGENT']) ? $_SERVER['HTTP_USER_AGENT'] : 'unknown';
+            $server_agent = $_SERVER['HTTP_USER_AGENT'] ?? 'unknown';
             $data = $server_agent . $config->get('security.salt', $this->grav['cache']->getKey());
             $this->rememberMe->setSalt(hash('sha512', $data));
 
@@ -519,25 +534,23 @@ class Login
     }
 
     /**
-     * @param User $user
-     * @param Page $page
+     * @param UserInterface $user
+     * @param PageInterface $page
      * @param Data|null $config
      * @return bool
      */
-    public function isUserAuthorizedForPage(User $user, Page $page, $config = null)
+    public function isUserAuthorizedForPage(UserInterface $user, PageInterface $page, $config = null)
     {
         $header = $page->header();
         $rules = isset($header->access) ? (array)$header->access : [];
 
-        if ($config !== null && $config->get('parent_acl')) {
+        if (!$rules && $config !== null && $config->get('parent_acl')) {
             // If page has no ACL rules, use its parent's rules
-            if (!$rules) {
-                $parent = $page->parent();
-                while (!$rules and $parent) {
-                    $header = $parent->header();
-                    $rules = isset($header->access) ? (array)$header->access : [];
-                    $parent = $parent->parent();
-                }
+            $parent = $page->parent();
+            while (!$rules and $parent) {
+                $header = $parent->header();
+                $rules = isset($header->access) ? (array)$header->access : [];
+                $parent = $parent->parent();
             }
         }
 
@@ -552,16 +565,14 @@ class Login
 
         // Continue to the page if user is authorized to access the page.
         foreach ($rules as $rule => $value) {
-            if (is_array($value)) {
+            if (\is_array($value)) {
                 foreach ($value as $nested_rule => $nested_value) {
-                    if ($user->authorize($rule . '.' . $nested_rule) == $nested_value) {
+                    if ($user->authorize($rule . '.' . $nested_rule) === Utils::isPositive($nested_value)) {
                         return true;
                     }
                 }
-            } else {
-                if ($user->authorize($rule) == $value) {
-                    return true;
-                }
+            } elseif ($user->authorize($rule) === Utils::isPositive($value)) {
+                return true;
             }
         }
 
@@ -571,14 +582,14 @@ class Login
     /**
      * Check if user may use password reset functionality.
      *
-     * @param User   $user
-     * @param string $field
-     * @param int    $count
-     * @param int    $interval
+     * @param UserInterface $user
+     * @param string        $field
+     * @param int           $count
+     * @param int           $interval
      * @return bool
      * @deprecated 2.5.0 Use $grav['login']->getRateLimiter($context) instead. See Grav\Plugin\Login\RateLimiter class.
      */
-    public function isUserRateLimited(User $user, $field, $count, $interval)
+    public function isUserRateLimited(UserInterface $user, $field, $count, $interval)
     {
         if ($count > 0) {
             if (!isset($user->{$field})) {
@@ -592,7 +603,7 @@ class Login
                 }
             }
 
-            if (count($actual_resets) >= $count) {
+            if (\count($actual_resets) >= $count) {
                 return true;
             }
             $actual_resets[] = time(); // current reset
@@ -605,11 +616,11 @@ class Login
     /**
      * Reset the rate limit counter.
      *
-     * @param User   $user
-     * @param string $field
+     * @param UserInterface $user
+     * @param string        $field
      * @deprecated 2.5.0 Use $grav['login']->getRateLimiter($context) instead. See Grav\Plugin\Login\RateLimiter class.
      */
-    public function resetRateLimit(User $user, $field)
+    public function resetRateLimit(UserInterface $user, $field)
     {
         $user->{$field} = [];
     }
@@ -617,12 +628,12 @@ class Login
     /**
      * Get Current logged in user
      *
-     * @return User
+     * @return UserInterface
      * @deprecated 2.5.0 Use $grav['user'] instead.
      */
     public function getUser()
     {
-        /** @var User $user */
+        /** @var UserInterface $user */
         return $this->grav['user'];
     }
 
@@ -633,9 +644,7 @@ class Login
 
     public function getProviderLoginTemplates()
     {
-        $templates =  $this->provider_login_templates;
-
-        return $templates;
+        return $this->provider_login_templates;
     }
 
 }
