@@ -1,9 +1,10 @@
 <?php
+
 namespace Grav\Plugin\Admin;
 
 use Grav\Common\Config\Config;
 use Grav\Common\Grav;
-use Grav\Common\Page\Page;
+use Grav\Common\Page\Interfaces\PageInterface;
 
 /**
  * Class Popularity
@@ -51,7 +52,12 @@ class Popularity
             return;
         }
 
-        /** @var Page $page */
+        // Respect visitors "do not track" setting
+        if (!Grav::instance()['browser']->isTrackable()) {
+            return;
+        }
+
+        /** @var PageInterface $page */
         $page         = Grav::instance()['page'];
         $relative_url = str_replace(Grav::instance()['base_url_relative'], '', $page->url());
 
@@ -123,8 +129,10 @@ class Popularity
         $labels = [];
         $data   = [];
 
+        /** @var Admin $admin */
+        $admin = Grav::instance()['admin'];
         foreach ($chart_data as $date => $count) {
-            $labels[] = Grav::instance()['grav']['admin']->translate([
+            $labels[] = $admin::translate([
                 'PLUGIN_ADMIN.' . strtoupper(date('D', strtotime($date)))]) .
                 '<br>' . date('M d', strtotime($date));
             $data[]   = $count;
