@@ -106,6 +106,7 @@ export default class FilesField {
         if (!this.container.length) { return; }
 
         this.urls = {};
+        this.customPost = this.container.data('filePostAdd') || {};
         this.options = Object.assign({}, Dictionary, DropzoneMediaConfig, {
             klass: this,
             url: this.container.data('file-url-add') || config.current_url,
@@ -133,6 +134,7 @@ export default class FilesField {
             const URL = Object.keys(value).filter((key) => value[key].name === filename).shift();
             target.attr('href', `${config.base_url_simple}/${URL}`);
         });
+
     }
 
     initDropzone() {
@@ -162,6 +164,7 @@ export default class FilesField {
 
             file.remove();
         });
+
     }
 
     getURI() {
@@ -169,10 +172,17 @@ export default class FilesField {
     }
 
     onDropzoneSending(file, xhr, formData) {
-        formData.append('name', this.options.dotNotation);
+        if (Object.keys(this.customPost).length) {
+            Object.keys(this.customPost).forEach((key) => {
+                formData.append(key, this.customPost[key]);
+            });
+        } else {
+            formData.append('name', this.options.dotNotation);
+            formData.append('task', 'filesupload');
+            formData.append('uri', this.getURI());
+        }
+
         formData.append('admin-nonce', config.admin_nonce);
-        formData.append('task', 'filesupload');
-        formData.append('uri', this.getURI());
     }
 
     onDropzoneSuccess(file, response, xhr) {
@@ -229,7 +239,7 @@ export default class FilesField {
 
     onDropzoneRemovedFile(file, ...extra) {
         if (!file.accepted || file.rejected) { return; }
-        let url = file.removeUrl || this.urls.delete;
+        let url = file.removeUrl || this.urls.delete || this.options.url;
         let path = (url || '').match(/path:(.*)\//);
         let body = { filename: file.name, uri: this.getURI() };
 
@@ -237,6 +247,17 @@ export default class FilesField {
             body.task = 'filessessionremove';
             body.session = file.sessionParams;
         }
+
+        const customPost = this.container.data('filePostRemove') || {};
+        if (Object.keys(customPost).length) {
+            body = {};
+            Object.keys(customPost).forEach((key) => {
+                body[key] = customPost[key];
+            });
+        }
+
+        body['filename'] = file.name;
+        body['admin-nonce'] = config.admin_nonce;
 
         request(url, { method: 'post', body }, () => {
             if (!path) { return; }
