@@ -61,7 +61,7 @@ if ($('.thumbs-list-container .dropzone')[0]) {
         $('.dropzone')[0].dropzone.files.forEach(function(file) { file.previewElement.remove(); });
         $('.dropzone').first().removeClass('dz-started');
 
-        request(config.base_url_relative + '/media-manager.json/task:clearMediaCache', { method: 'post', body }, () => {
+        request(`${config.base_url_relative}/media-manager.json/task${config.param_sep}clearMediaCache`, { method: 'post', body }, () => {
             filterFiles();
         });
     });
@@ -69,16 +69,16 @@ if ($('.thumbs-list-container .dropzone')[0]) {
 
 /* handle loading media */
 var loadMedia = function loadMedia(filters, callback) {
-    var url = config.base_url_relative + '/media.json/tmpl:media-list-content/index:' + global_index;
+    var url = `${config.base_url_relative}/media.json/tmpl${config.param_sep}media-list-content/index${config.param_sep}${global_index}`;
 
     if (filters.page) {
-        url += '/page:' + (filters.page).split('/').join('%5C');
+        url += `/page${config.param_sep}${(filters.page).split('/').join('%5C')}`;
     }
     if (filters.type && filters.type !== '*') {
-        url += '/type:' + filters.type;
+        url += `/type${config.param_sep}${filters.type}`;
     }
     if (filters.date && filters.date !== '*') {
-        url += '/date:' + filters.date;
+        url += `/date${config.param_sep}${filters.date}`;
     }
 
     if (!isLoading) {
