@@ -138,7 +138,11 @@ class FeedPlugin extends Plugin
 
         /** @var Data\Blueprint $blueprint */
         $blueprint = $event['blueprint'];
-        if (!$inEvent && $blueprint->name == 'blog_list') {
+        $form = $blueprint->form();
+
+        $blog_tab_exists = isset($form['fields']['tabs']['fields']['blog']);
+
+        if (!$inEvent && $blog_tab_exists) {
             $inEvent = true;
             $blueprints = new Data\Blueprints(__DIR__ . '/blueprints/');
             $extends = $blueprints->get('feed');
