@@ -67,7 +67,8 @@ export default class FilesField {
                 status: Dropzone.ADDED,
                 accepted: true,
                 url: this.options.url,
-                removeUrl: data.remove
+                removeUrl: data.remove,
+                data
             };
 
             dropzone.files.push(mock);
@@ -83,7 +84,12 @@ export default class FilesField {
     }
 
     onDropzoneSending(file, xhr, formData) {
-        formData.append('__form-name__', this.container.closest('form').find('[name="__form-name__"]').val());
+        const form = this.container.closest('form');
+        const unique_id = form.find('[name="__unique_form_id__"]');
+        formData.append('__form-name__', form.find('[name="__form-name__"]').val());
+        if (unique_id.length) {
+            formData.append('__unique_form_id__', unique_id.val());
+        }
         formData.append('__form-file-uploader__', 1);
         formData.append('name', this.options.dotNotation);
         formData.append('form-nonce', config.form_nonce);
@@ -139,12 +145,14 @@ export default class FilesField {
 
     onDropzoneRemovedFile(file, ...extra) {
         if (!file.accepted || file.rejected) { return; }
+        const form = this.container.closest('form');
+        const unique_id = form.find('[name="__unique_form_id__"]');
         let url = file.removeUrl || this.urls.delete || `${location.href}.json`;
         let path = (url || '').match(/path:(.*)\//);
         let data = new FormData();
 
         data.append('filename', file.name);
-        data.append('__form-name__', this.container.closest('form').find('[name="__form-name__"]').val());
+        data.append('__form-name__', form.find('[name="__form-name__"]').val());
         data.append('name', this.options.dotNotation);
         data.append('form-nonce', config.form_nonce);
         data.append('uri', this.getURI());
@@ -152,6 +160,10 @@ export default class FilesField {
         if (file.sessionParams) {
             data.append('__form-file-remover__', '1');
             data.append('session', file.sessionParams);
+        }
+
+        if (unique_id.length) {
+            data.append('__unique_form_id__', unique_id.val());
         }
 
         $.ajax({
@@ -215,6 +227,7 @@ export default class FilesField {
     }
 }
 
+/*
 export function UriToMarkdown(uri) {
     uri = uri.replace(/@3x|@2x|@1x/, '');
     uri = uri.replace(/\(/g, '%28');
@@ -222,6 +235,7 @@ export function UriToMarkdown(uri) {
 
     return uri.match(/\.(jpe?g|png|gif|svg)$/i) ? `![](${uri})` : `[${decodeURI(uri)}](${uri})`;
 }
+*/
 
 let instances = [];
 let cache = $();
@@ -258,10 +272,10 @@ const addNode = (container) => {
         resizeQuality: settings.resizeQuality || null,
         accept: function(file, done) {
             const resolution = settings.resolution;
+            let error = '';
             if (!resolution) return done();
 
             setTimeout(() => {
-                let error = '';
                 if (resolution.min) {
                     Object.keys(resolution.min).forEach((attr) => {
                         if (file[attr] < resolution.min[attr]) {
