@@ -1,4 +1,7 @@
+import './form';
 import FileInstances from './file';
 import ArrayInstances from './array';
+import PageMedia, { Instance as PageMediaInstances } from './media';
+import './tabs';
 
-export default { FileInstances, ArrayInstances };
+export default { FileInstances, ArrayInstances, Media: { PageMedia, PageMediaInstances } };
