@@ -1,10 +1,12 @@
 <?php
+
 /**
  * @package    Grav\Plugin\Login
  *
  * @copyright  Copyright (C) 2014 - 2017 RocketTheme, LLC. All rights reserved.
  * @license    MIT License; see LICENSE file for details.
  */
+
 namespace Grav\Plugin\Login;
 
 /**
@@ -56,7 +58,7 @@ class RateLimiter
             return false;
         }
 
-        return $this->maxCount && count($this->getAttempts($key, $type)) > $this->maxCount;
+        return $this->maxCount && \count($this->getAttempts($key, $type)) > $this->maxCount;
     }
 
     /**
