@@ -15,25 +15,29 @@ var paths = {
     source: src_dir
 };
 
-gulp.task('watch', function() {
-  gulp.watch(watch_dir, ['build']);
-});
+function watch() {
+  return gulp.watch(watch_dir, build);
+}
 
-gulp.task('build', function() {
-  gulp.src(paths.source)
-    .pipe(sourcemaps.init())
-    .pipe(sass({outputStyle: 'compact', precision: 10})
-      .on('error', sass.logError)
-    )
-    .pipe(sourcemaps.write())
-    .pipe(autoprefixer())       
-    .pipe(gulp.dest(dest_dir))
-    .pipe(csscomb())
-    .pipe(cleancss())
-    .pipe(rename({
-      suffix: '.min'
-    }))
-    .pipe(gulp.dest(dest_dir));
-});
+function build() {
+  return gulp.src(paths.source)
+      .pipe(sourcemaps.init())
+      .pipe(sass({
+            outputStyle: 'compact',
+            precision: 10
+          }).on('error', sass.logError)
+      )
+      .pipe(sourcemaps.write())
+      .pipe(autoprefixer())
+      .pipe(gulp.dest(dest_dir))
+      .pipe(csscomb())
+      .pipe(cleancss())
+      .pipe(rename({
+        suffix: '.min'
+      }))
+      .pipe(gulp.dest(dest_dir));
+}
 
-gulp.task('default', ['build']);
+exports.watch = watch;
+exports.build = build;
+exports.default = build;
